chore: delete unused file
