--- .kiro/specs/mise-task-runner-mcp/requirements.md
@@ -1,62 +1,0 @@
-# 要件ドキュメント
-
-## はじめに
-
-この機能は、miseのタスクランナーをModel Context Protocol（MCP）サーバーと統合し、ユーザーがMCPインターフェースを通じてmise.tomlファイルで定義されたタスクを選択して実行できるようにします。MCPサーバーはmise.tomlファイルを解析し、タスクの説明を抽出して、選択可能なオプションとして提供します。タスクが選択されると、ローカルで実行されます。このソリューションはDenoで構築され、npxを通じてインストール可能になります。
-
-## 要件
-
-### 要件1：MCPサーバーの実装
-
-**ユーザーストーリー:** 開発者として、miseのタスクランナーと連携できるMCPサーバーが欲しい。それによりMCPインターフェースを通じてmiseタスクを選択し実行できるようにしたい。
-
-#### 受け入れ基準
-
-1. MCPサーバーが起動されるとき、MCPクライアントとの接続を確立すること。
-2. MCPサーバーが利用可能なツールのリクエストを受け取るとき、miseタスクとその説明のリストを返すこと。
-3. MCPサーバーが初期化されるとき、mise.tomlファイルを解析してタスク情報を抽出できること。
-4. MCPサーバーが実行中のとき、MCPプロトコル仕様との互換性を維持すること。
-
-### 要件2：タスクの検出とリスト表示
-
-**ユーザーストーリー:** 開発者として、利用可能なmiseタスクとその説明のリストを見たい。それにより適切なタスクを簡単に識別して選択できるようにしたい。
-
-#### 受け入れ基準
-
-1. MCPクライアントが利用可能なツールをリクエストするとき、サーバーはmiseタスクの名前と説明のリストを返すこと。
-2. mise.tomlファイルにタスク定義が含まれるとき、サーバーはタスク名とその説明の両方を抽出すること。
-3. ワークスペースに複数のmise.tomlファイルが存在するとき、サーバーはすべてのファイルからタスクを集約すること。
-4. mise.tomlファイルが更新されるとき、サーバーは次回のリクエスト時にタスクリストにこれらの変更を反映すること。
-
-### 要件3：タスクの実行
-
-**ユーザーストーリー:** 開発者として、選択したmiseタスクをローカルで実行したい。それによりMCPインターフェースを通じて開発ワークフローを実行できるようにしたい。
-
-#### 受け入れ基準
-
-1. タスクがMCPインターフェースを通じて選択されるとき、サーバーはmiseを使用してローカルで実行すること。
-2. タスク実行がリクエストされるとき、サーバーは実行出力をMCPクライアントに返すこと。
-3. タスク実行が失敗するとき、サーバーは適切なエラー情報をMCPクライアントに提供すること。
-4. タスクがパラメータを必要とするとき、サーバーはこれらのパラメータを受け入れてmiseタスクに渡すこと。
-
-### 要件4：インストールと配布
-
-**ユーザーストーリー:** 開発者として、npxを使用してmise MCPサーバーを簡単にインストールしたい。それにより統合を迅速にセットアップして使用できるようにしたい。
-
-#### 受け入れ基準
-
-1. ユーザーがnpxコマンドを実行するとき、mise MCPサーバーがインストールされ利用可能になること。
-2. サーバーがインストールされるとき、必要なすべての依存関係を含むこと。
-3. サーバーがインストールされるとき、明確な使用方法の説明を提供すること。
-4. サーバーがインストールされるとき、追加の設定手順なしで実行可能であること。
-
-### 要件5：ドキュメントとヘルプ
-
-**ユーザーストーリー:** 開発者として、mise MCPサーバーの包括的なドキュメントが欲しい。それにより効果的に使用および設定する方法を理解できるようにしたい。
-
-#### 受け入れ基準
-
-1. サーバーがインストールされるとき、コマンドラインフラグを通じてヘルプドキュメントを提供すること。
-2. ユーザーがヘルプをリクエストするとき、サーバーは使用方法と利用可能なオプションを表示すること。
-3. サーバーが配布されるとき、セットアップと使用例を含むREADMEドキュメントを含むこと。
-4. エラーが発生するとき、サーバーは潜在的な解決策を含む明確なエラーメッセージを提供すること。
