--- .kiro/specs/mise-task-runner-mcp/tasks.md
@@ -1,100 +1,0 @@
-# 実装計画
-
-- [ ] 1. プロジェクト構造とコアインターフェースのセットアップ
-  - プロジェクトディレクトリ構造を作成
-  - 基本的な依存関係を設定
-  - コアインターフェースと型定義を実装
-  - _要件: 1.1, 1.3, 1.4_
-
-- [ ] 2. Miseタスクパーサーの実装
-  - [ ] 2.1 mise.tomlファイル検索機能の実装
-    - ワークスペース内のmise.tomlファイルを再帰的に検索する機能を実装
-    - 検索結果をキャッシュする仕組みを追加
-    - _要件: 2.1, 2.3, 2.4_
-
-  - [ ] 2.2 TOMLパーサーの実装
-    - mise.tomlファイルを解析するパーサーを実装
-    - タスクセクションの抽出ロジックを実装
-    - _要件: 2.2, 2.3_
-
-  - [ ] 2.3 タスク情報抽出機能の実装
-    - タスク名と説明を抽出する機能を実装
-    - タスクの依存関係を解析する機能を実装
-    - _要件: 2.1, 2.2_
-
-- [ ] 3. MCPサーバーの基本実装
-  - [ ] 3.1 MCPプロトコルの実装
-    - MCPプロトコルに準拠した基本的なサーバー構造を実装
-    - リクエスト/レスポンスハンドラーを実装
-    - _要件: 1.1, 1.4_
-
-  - [ ] 3.2 ツールリスト機能の実装
-    - Miseタスクをツールとして提供する機能を実装
-    - タスク説明をツール説明として使用する機能を実装
-    - _要件: 1.2, 2.1_
-
-- [ ] 4. タスク実行エンジンの実装
-  - [ ] 4.1 miseコマンド実行機能の実装
-    - miseコマンドを構築して実行する機能を実装
-    - コマンド出力をキャプチャする機能を実装
-    - _要件: 3.1, 3.2_
-
-  - [ ] 4.2 パラメータ処理機能の実装
-    - タスクパラメータを処理する機能を実装
-    - パラメータをmiseコマンドに渡す機能を実装
-    - _要件: 3.4_
-
-  - [ ] 4.3 エラーハンドリングの実装
-    - タスク実行エラーを検出して処理する機能を実装
-    - エラー情報をクライアントに返す機能を実装
-    - _要件: 3.3_
-
-- [ ] 5. CLIの実装
-  - [ ] 5.1 コマンドライン引数処理の実装
-    - コマンドライン引数を解析する機能を実装
-    - ヘルプとバージョン情報を表示する機能を実装
-    - _要件: 5.1, 5.2_
-
-  - [ ] 5.2 サーバー起動/停止機能の実装
-    - サーバーを起動する機能を実装
-    - サーバーを適切に停止する機能を実装
-    - _要件: 1.1_
-
-- [ ] 6. npxインストール対応
-  - [ ] 6.1 npmパッケージ設定の実装
-    - package.jsonの設定
-    - bin設定の実装
-    - _要件: 4.1, 4.2_
-
-  - [ ] 6.2 インストールスクリプトの実装
-    - 依存関係の自動インストール機能を実装
-    - 初期設定を行う機能を実装
-    - _要件: 4.2, 4.4_
-
-- [ ] 7. テストの実装
-  - [ ] 7.1 ユニットテストの実装
-    - 各コンポーネントのユニットテストを実装
-    - モックとスタブを使用したテスト環境を構築
-    - _要件: 1.3, 2.2, 3.1_
-
-  - [ ] 7.2 統合テストの実装
-    - コンポーネント間の連携をテストする統合テストを実装
-    - サンプルmise.tomlを使用したテストケースを実装
-    - _要件: 1.2, 2.1, 3.1_
-
-- [ ] 8. ドキュメントの作成
-  - [ ] 8.1 READMEの作成
-    - インストール方法の説明
-    - 使用方法の説明と例
-    - トラブルシューティングガイド
-    - _要件: 5.3_
-
-  - [ ] 8.2 APIドキュメントの作成
-    - MCPツールインターフェースの説明
-    - パラメータと戻り値の説明
-    - _要件: 5.3_
-
-  - [ ] 8.3 エラーメッセージの改善
-    - 明確なエラーメッセージの実装
-    - 解決策の提案を含むエラーハンドリングの実装
-    - _要件: 5.4_
