+++ doc/product.md
@@ -1,0 +1,12 @@
+# Product Overview
+
+mise-task-mcp-server is an MCP (Model Context Protocol) server that enables AI assistants to interact with mise tasks. It provides a bridge between AI tools and the mise task runner, allowing assistants to list and execute mise tasks from projects.
+
+## Key Features
+- List available mise tasks from mise.toml configuration
+- Execute mise tasks programmatically
+- Cross-platform binary distribution (Linux, macOS, Windows)
+- Stdio-based MCP server implementation
+
+## Target Use Case
+Enables AI assistants to understand and interact with mise-based project workflows, making it easier to automate development tasks and provide contextual assistance for mise-managed projects.
