--- mcp-server/mcp_server_sdk.ts
@@ -1,291 +1,0 @@
-import { Server } from "@modelcontextprotocol/sdk/server/index.js";
-import { StdioServerTransport } from "@modelcontextprotocol/sdk/server/stdio.js";
-import {
-  ListToolsRequestSchema,
-  CallToolRequestSchema,
-} from "@modelcontextprotocol/sdk/types.js";
-import { parse } from "@std/toml";
-
-interface Task {
-  name: string;
-  description?: string;
-  dependencies?: string[];
-}
-
-interface MiseToml {
-  tasks?: {
-    [key: string]: {
-      run?: string;
-      description?: string;
-      depends?: string[];
-    };
-  };
-}
-
-export class MiseMCPServer {
-  private server: Server;
-
-  constructor() {
-    this.server = new Server(
-      {
-        name: "mise-task-server",
-        version: "0.2.0",
-      },
-      {
-        capabilities: {
-          tools: {},
-        },
-      }
-    );
-
-    this.setupHandlers();
-  }
-
-  private setupHandlers() {
-    // Handle tool listing
-    this.server.setRequestHandler(ListToolsRequestSchema, async () => {
-      return await this.getTools();
-    });
-
-    // Handle tool calls
-    this.server.setRequestHandler(CallToolRequestSchema, async (request) => {
-      return await this.handleToolCall(request.params.name, request.params.arguments);
-    });
-  }
-
-  async listTasks(directory?: string): Promise<Task[]> {
-    const searchDir = directory || Deno.cwd();
-    const tasks: Task[] = [];
-
-    // Look for mise.toml file
-    const miseTomlPath = `${searchDir}/mise.toml`;
-    try {
-      const tomlContent = await Deno.readTextFile(miseTomlPath);
-      const parsedTasks = await this.parseMiseTomlTasks(tomlContent);
-      tasks.push(...parsedTasks);
-    } catch {
-      // mise.toml not found or cannot be read
-    }
-
-    // Look for file tasks in mise-tasks directory
-    const miseTasksDir = `${searchDir}/mise-tasks`;
-    try {
-      for await (const entry of Deno.readDir(miseTasksDir)) {
-        if (entry.isFile) {
-          tasks.push({
-            name: entry.name,
-            description: `File task: ${entry.name}`,
-          });
-        }
-      }
-    } catch {
-      // mise-tasks directory not found
-    }
-
-    return tasks;
-  }
-
-  async parseMiseTomlTasks(tomlContent: string): Promise<Task[]> {
-    const parsed = parse(tomlContent) as MiseToml;
-    const tasks: Task[] = [];
-
-    if (parsed.tasks) {
-      for (const [name, config] of Object.entries(parsed.tasks)) {
-        tasks.push({
-          name,
-          description: config.description,
-          dependencies: config.depends,
-        });
-      }
-    }
-
-    return tasks;
-  }
-
-  async runTask(taskName: string, args: string[] = []): Promise<{
-    success: boolean;
-    output?: string;
-    error?: string;
-  }> {
-    try {
-      // For testing purposes, if taskName is 'echo' or 'nonexistent-command',
-      // run them directly. Otherwise, use mise to run the task
-      if (taskName === "echo" || taskName === "nonexistent-command") {
-        const command = new Deno.Command(taskName, {
-          args,
-          stdout: "piped",
-          stderr: "piped",
-        });
-
-        const { code, stdout, stderr } = await command.output();
-
-        const output = new TextDecoder().decode(stdout);
-        const error = new TextDecoder().decode(stderr);
-
-        return {
-          success: code === 0,
-          output: output || undefined,
-          error: code !== 0 ? error || "Task failed" : undefined,
-        };
-      }
-
-      // Use mise to run the task
-      const command = new Deno.Command("mise", {
-        args: ["run", taskName, ...args],
-        stdout: "piped",
-        stderr: "piped",
-      });
-
-      const { code, stdout, stderr } = await command.output();
-
-      const output = new TextDecoder().decode(stdout);
-      const error = new TextDecoder().decode(stderr);
-
-      return {
-        success: code === 0,
-        output: output || undefined,
-        error: code !== 0 ? error || "Task failed" : undefined,
-      };
-    } catch (error) {
-      return {
-        success: false,
-        error: error instanceof Error ? error.message : String(error),
-      };
-    }
-  }
-
-  async start() {
-    const transport = new StdioServerTransport();
-    await this.server.connect(transport);
-  }
-
-  async getTools() {
-    const tasks = await this.listTasks();
-    
-    return {
-      tools: tasks.map(task => this.createToolFromTask(task)),
-    };
-  }
-
-  private createToolFromTask(task: Task) {
-    return {
-      name: task.name,
-      description: task.description || `Run ${task.name} task`,
-      inputSchema: this.createTaskInputSchema(),
-    };
-  }
-
-  private createTaskInputSchema() {
-    return {
-      type: "object",
-      properties: {
-        args: {
-          type: "array",
-          items: {
-            type: "string",
-          },
-          description: "Arguments to pass to the task",
-        },
-      },
-    };
-  }
-
-  private async handleToolCall(toolName: string, args: any) {
-    try {
-      // Handle legacy tools for backward compatibility
-      if (toolName === "mise-list-tasks") {
-        return await this.handleListTasksTool(args?.directory);
-      }
-
-      if (toolName === "mise-run-task") {
-        return await this.handleRunTaskTool(args);
-      }
-
-      // Handle dynamic tools (mise.toml tasks)
-      return await this.handleDynamicTool(toolName, args);
-    } catch (error) {
-      return {
-        content: [
-          {
-            type: "text",
-            text: `Error: ${error instanceof Error ? error.message : "Unknown error occurred"}`,
-          },
-        ],
-      };
-    }
-  }
-
-  private async handleListTasksTool(directory?: string) {
-    const tasks = await this.listTasks(directory);
-    
-    return {
-      content: [
-        {
-          type: "text",
-          text: JSON.stringify(tasks, null, 2),
-        },
-      ],
-    };
-  }
-
-  private async handleRunTaskTool(args: any) {
-    const { task, args: taskArgs } = args as {
-      task: string;
-      args?: string[];
-    };
-    
-    if (!task || typeof task !== "string") {
-      throw new Error("Task name is required and must be a string");
-    }
-
-    const result = await this.runTask(task, taskArgs || []);
-    
-    return {
-      content: [
-        {
-          type: "text",
-          text: result.success ? result.output || "Task completed successfully" : result.error || "Task failed",
-        },
-      ],
-    };
-  }
-
-  private async handleDynamicTool(toolName: string, args: any) {
-    if (!toolName || typeof toolName !== "string") {
-      throw new Error("Tool name is required and must be a string");
-    }
-
-    const taskArgs = args?.args || [];
-    if (taskArgs && !Array.isArray(taskArgs)) {
-      throw new Error("Arguments must be an array");
-    }
-
-    const result = await this.runTask(toolName, taskArgs);
-    
-    return {
-      content: [
-        {
-          type: "text",
-          text: result.success ? result.output || "Task completed successfully" : result.error || "Task failed",
-        },
-      ],
-    };
-  }
-
-  async callTool(request: { name: string; arguments: any }) {
-    const result = await this.runTask(request.name, request.arguments?.args || []);
-    
-    return {
-      success: result.success,
-      content: [
-        {
-          type: "text",
-          text: result.success ? result.output || "Task completed successfully" : result.error || "Task failed",
-        },
-      ],
-    };
-  }
-}
-
-// For backward compatibility
-export { MiseMCPServer as MCPServer };
